requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.6
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-import-export==2.5.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.8.3
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
kaleido==0.2.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.4.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.3
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
